Use an enum attribute to define the plan levels and set a default
Why:
When a `company` is created, this `enum` attribute will limit the plan levels to the 5 defined. If one is passed that isn't defined, an `ArgumentError` is passed. If a plan level isn't specified, it is assumed the company is in a trial period and will be assigned the `basic` plan level.